att verificação usuario nas features notas

+++ src/app/features/note/usecases/delete-note.usecase.ts
@@ -1,8 +1,19 @@
 import { CacheRepository } from "../../../shared/repositories/cache.repository";
+import { UserRepository } from "../../user/repository/user.repository";
 import { NoteRepository } from "../repository/note.repository";
 
 export class DeleteNoteUsecase {
   public async execute(id: string, idUser: string) {
+    const userRepository = new UserRepository();
+    const user = await userRepository.get(idUser);
+
+    if (!user) {
+      return {
+        ok: false,
+        code: 404,
+        message: "Usuário não encontrado",
+      };
+    }
     const repository = new NoteRepository();
     const result = await repository.delete(id);
